app/requirements.txt
grpcio==1.32.0
gunicorn==20.0.4
Flask==0.10.1
Jinja2==2.7.1
h5py==2.10.0
imutils==0.5.4
ipykernel==5.4.3
jedi==0.18.0
joblib==1.0.0
Keras-Preprocessing==1.1.2
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.1.5
parso==0.8.1
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
ruamel-yaml==0.15.87
scikit-learn==0.24.1
scipy==1.5.4
sklearn==0.0
wheel==0.35.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tornado==6.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.3
